taskduration changed to float

+++ Backend/Data/Entitites/TaskItem.cs
@@ -4,11 +4,11 @@ public class TaskItem
 {
     public int Id { get; set; }
     public string Name { get; set; } = "";
-    public int DurationHours { get; set; }
+    public float DurationHours { get; set; }
 
     public ICollection<TaskQualification> RequiredQualifications { get; set; } = [];
     public ICollection<TaskTool> RequiredTools { get; set; } = [];
-    public TaskItem(string name, int durationHours)
+    public TaskItem(string name, float durationHours)
     {
         Name = name;
         DurationHours = durationHours;
